Dashboard tabs, dashboard groups, layout updates

+++ src/kibana/plugins/visualize/saved_visualizations/_saved_vis.js
@@ -46,7 +46,7 @@ define(function (require) {
       visState: 'json',
       description: 'string',
       savedSearchId: 'string',
-      version: 'integer'
+      version: 'long'
     };
 
     SavedVis.searchSource = true;
